renamed stream_adaptor
